add a POC for a stateful trigger of the help buoy
This change adds a new property on application state which indicates
whether the user has exited the intro.js "tour" on the Catalog
page. The help buoy in /Components/Ui/Help.js is now a connected
component listening to state, and has an animated slide in which is
triggered by the tour ending.

Additionally, this change adds a custom stylesheet to override the
intro.js max-width, which was causing the background on the wider
intro boxes to fail to cover the full width of the box.

Note also that these diffs are, in some cases, mostly
reformatting. That is due to using prettier.js for all touched files.