move Help module from /UI to /Help
This change updates the Help.js module, moving it out of the UI
directory and into its own, to reflect it as a top-lever concern of
the app.

Additionally, several behaviors have changed:

The help component now responds to a boolean prop "showHelp" that
indicates whether the help button should be shown. This also allows
the app to trigger the showing of the botton at specific times.

The help button enters the page expanded, to further draw attention to
it as an interactable menu. The menu will close automatically after 3
seconds.

Because it now depends on an injected property from its parent,
Help.js is no longer a connected component.

This change includes updates to most imports of Help.js, but not
all (see subsequent commit).