add "Persist" service for local storage
This change adds a service to wrap local storage, providing a fallback
if it is not available.

Additionally:

add a constants.js file in the src root (this moves away from the
"enums" pattern)

add a module that provides inital application state for intros and
hints; this is consumed during the store init to set initial state for
redux.