add Intro component to control stepped tours
This change extracts the implementation of the intro.js Steps
component to a wrapping component, which handles integration with
local storage and redux (and is also where the custom css is
imported).

The steps config is passed as a prop. The component enables the intro
based on redux. When the intro exits, this component updates local
storage as well as application state so that the intro is not
triggered on subsequent page visits.

Associated redux changes are included in this change, discarding the
initial POC implementation. This change does not include this new
behavior on the Catalog page.