enable Intro component on Catalog page
This change replaces the initial inclusion of the intro.js Steps
component with the stateful Intro component.

With this change, visitors to the page will initially be presented
with an introduction, but once they exit the intro, a flag is set in
both application state and local storage, preventing the tour from
showing on subsequent page visits.