fix persistence bug and change intro control to toggle
This change corrects the two uses of the persistence service, applying
a JSON.parse to the stored value, because local storage stores falues
as strings. It is not plausible to handle this in the service itself,
because it has no knowledge of the type of value the caller is
expecting.

Additionally, this change converts the one-way "disable" control for
intros to a toggle.