omnibus help changes
This change includes several related updates:

Move the registration of hint and intro with the persistence service
out of the ui components, and up into App.js. Instantiating multiple
Hint components (.e.g. in CatalogSearch and SearchResults) may be
conflicting.

Fix a bug where navigating away from a page where the Tour is
active (unmounting the Intro component) triggers the onTourExit
handler, re-enabling it when the user navigates back to the page.

Detect active route for page-specific Intro, Hints, and Video. Add a
helper to map from pathName to pageName. (Note, in a future change, we
should make this a part of a Page component, with a canonical name,
which anchors all related app behavior, such as persistence, related
content, etc.)

Use the Help menu in the TopNavbar to control page-specific Hints and
to play a page-specific video. Add tooltips and icons to help menu items.

Rename the Help module to HelpFab, since it is no longer used and its
only responsibility is now the unused UI control for Hints. Remove it
from the Visualization page.

Override intro.js styling of hint beacons; replace with (?).

Minor changes: .gitignore; prettier reformatting for Catalog.js,
Visualization.js, and App.js.