several bug fixes
This change collects a few fixes prior to pushing Help changes to the
dev server in order to solicit feedback:

The changes to the route definitions in App.js broke the
/catalog/datasets/:id and /catalog/cruises/:id routes, this change
reverts those changes for these two routes.

Added a hint for dataset downloads

Updated text in the keyword search hint

Update the video on the catalog page

Updated style in NavControls to limit the video's modal window
stretching far beyond the video itself, and to make the excess marign
less ugly (this is a stopgap until we actually figure out how to made
the modal adaptive)

Make the modal close if user interacts with the help menu.