WIP implement hints
this change adds a wip custom implementation of hints

original approach was to try to make a hist a popover of a popover (in
material ui components) but that proved unworkable; this approach is a
components that wraps another component and absolutely positions a
beacon in a designated corner in proximity to the wrapped component.

downsides to this approach are that we need to make lots of little
style adjustments to get the layout flow to work well, on a component
by component basis (by contrast, the MUI approach calculates a
coordinate in the viewport and renders the hint in a portal, i.e., a
overlay layer that is outside the component tree of the component it
is anchored to)

this may become unwieldy in with adaptive modes on smaller screens.

additionally, the hint itself (what appears when you click on the
beacon) is currently a MUI Tooltip, and this has proved to exibit the
same positioning problems, and will need to be replaced with another
component layer