fixes for Intro
These changes refine the behavior of Introjs, including a filter that
removes any intro steps whose achoring elements are not selectable,
or out of the viewbox. Note also the redux changes that allow the
dispatcher of toggleIntro to specify a value, and the related
changes to persistIntroState in the payloadToValue it provides to the
persistence service.