homepage redesign
This change replaces the webflow landing page and About page with a
react-based homepage, a new design, and a site-wide navigation bar
update, as well as other small changes. The new design includes a News
feature, and a means for admins to manage news items, and a Gallery
page.

This change depends upon cmap-node-api changes to routing, both for
serving the react app for the root route, and for the necessary api
calls for the news service.

Additionally, a new set of common component, based on the new styles,
have been developed and included to ease future build out of new pages.

User api calls are now directed to /api

Site theme has been extracted out of the App.js module

Old assets for the webflow landing page have been deleted to reduce
deployment bundle size by ~50mb

Move admin menu for admin data submission dashboard out of the data
submission menu, into the user menu.

Fix missing timezone conversion in validator: When datasheets are
submitted with a time field formatted as a string, rather than as a
date or number, the formatDataSheet was failing to account for
automatically converted timezone when parsing the string as a date: we
assume that the submitted time is GMT, but the application interprets
it as a local time, which was then incorrectly converted back to GMT
when converted to an ISO string. This change corrects that behavion by
detecting what the local timezone offset is, and subtracting it from
the submitted time value.