validator v2 and dataset sample visualizations
An overhaul of the validator, primarily aimed at introducing
name-conflict resolution, but with a streamlined and informative UX,
the ability to choose which existing submission to update, additional
audits and fixes to auditing bugs--especially in handling time
values. This work includes a partial refactor of the validation tool,
moving the audit state to redux, and extracting many isolable concerns
into separate modules. The workbook audits themselves now constitute
their own module, adhere to a common interface, and are profiled for
performance.

The dataset detail page now displays a sample visualization of an
automatically selected (visualizable) variable, and grants to ability
to select among all visualizable variables and see a sample
visualization. Additionally the page style has been updated in certain
respects, including a more informative display of related cruises.